Dummy engine is now in dummy plugins.
